move code to the right files, fix linter issues
